# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/admin/plans/all": {
      "get": {
        "operationId": "AdminPlansController_getPlans",
        "summary": "Get all plans",
        "description": "Retrieve a paginated list of all plans with their prices and features",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page (default: 10)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number for pagination (default: 1)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search in plan name and description",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "description": "Filter by active status",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Plans retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "number"
                        },
                        "page": {
                          "type": "number"
                        },
                        "limit": {
                          "type": "number"
                        },
                        "totalPages": {
                          "type": "number"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Admin Plan Management"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin/plans/active": {
      "get": {
        "operationId": "AdminPlansController_getActivePlansWithPrices",
        "summary": "Get active plans with prices",
        "description": "Retrieve all active plans with their associated prices and features",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Active plans retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      },
                      "description": {
                        "type": "string"
                      },
                      "prices": {
                        "type": "array",
                        "items": {
                          "type": "object"
                        }
                      },
                      "features": {
                        "type": "array",
                        "items": {
                          "type": "object"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Admin Plan Management"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin/plans/{id}": {
      "get": {
        "operationId": "AdminPlansController_getPlanById",
        "summary": "Get plan by ID",
        "description": "Retrieve detailed information about a specific plan including prices and features",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Plan ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Plan not found"
          }
        },
        "tags": [
          "Admin Plan Management"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "operationId": "AdminPlansController_updatePlan",
        "summary": "Update an existing plan",
        "description": "Update plan properties such as name, description, or active status",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Plan ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Plan update data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlanDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Plan updated successfully"
          },
          "400": {
            "description": "Invalid input data"
          },
          "404": {
            "description": "Plan not found"
          },
          "409": {
            "description": "Plan with this slug already exists for this location"
          }
        },
        "tags": [
          "Admin Plan Management"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "AdminPlansController_deletePlan",
        "summary": "Delete a plan",
        "description": "Soft delete a plan and all its associated features and prices",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Plan ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Plan deleted successfully"
          },
          "400": {
            "description": "Failed to delete plan"
          },
          "404": {
            "description": "Plan not found"
          }
        },
        "tags": [
          "Admin Plan Management"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/admin/plans/plan-prices/{id}": {
      "get": {
        "operationId": "AdminPlansController_getPlanPriceById",
        "summary": "Get plan price by ID",
        "description": "Retrieve detailed information about a specific plan price",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Plan Price ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Plan price retrieved successfully"
          },
          "404": {
            "description": "Plan price not found"
          }
        },
        "tags": [
          "Admin Plan Management"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "operationId": "AdminPlansController_updatePlanPrice",
        "summary": "Update plan price",
        "description": "Update plan price properties such as amount, currency, or billing cycle",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Plan Price ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Plan price update data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlanPriceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Plan price updated successfully"
          },
          "400": {
            "description": "Invalid input data"
          },
          "404": {
            "description": "Plan price not found"
          }
        },
        "tags": [
          "Admin Plan Management"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "AdminPlansController_deletePlanPrice",
        "summary": "Delete plan price",
        "description": "Soft delete a plan price configuration",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Plan Price ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Plan price deleted successfully"
          },
          "400": {
            "description": "Failed to delete plan price"
          },
          "404": {
            "description": "Plan price not found"
          }
        },
        "tags": [
          "Admin Plan Management"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/plans/all": {
      "get": {
        "operationId": "PlansController_getPlans",
        "summary": "Get all plans",
        "description": "Retrieve a paginated list of all plans with their prices and features",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page (default: 10)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number for pagination (default: 1)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search in plan name and description",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "description": "Filter by active status",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Plans retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "number"
                        },
                        "page": {
                          "type": "number"
                        },
                        "limit": {
                          "type": "number"
                        },
                        "totalPages": {
                          "type": "number"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Plan Management"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/plans/active": {
      "get": {
        "operationId": "PlansController_getActivePlansWithPrices",
        "summary": "Get active plans with prices",
        "description": "Retrieve all active plans with their associated prices and features",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Active plans retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      },
                      "description": {
                        "type": "string"
                      },
                      "prices": {
                        "type": "array",
                        "items": {
                          "type": "object"
                        }
                      },
                      "features": {
                        "type": "array",
                        "items": {
                          "type": "object"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Plan Management"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/plans/{id}": {
      "get": {
        "operationId": "PlansController_getPlanById",
        "summary": "Get plan by ID",
        "description": "Retrieve detailed information about a specific plan including prices and features",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Plan ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Plan retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Plan not found"
          }
        },
        "tags": [
          "Plan Management"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/plans/plan-prices/{id}": {
      "get": {
        "operationId": "PlansController_getPlanPriceById",
        "summary": "Get plan price by ID",
        "description": "Retrieve detailed information about a specific plan price",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Plan Price ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Plan price retrieved successfully"
          },
          "404": {
            "description": "Plan price not found"
          }
        },
        "tags": [
          "Plan Management"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Virtual Mail Hub",
    "description": "The Virtual Mail Hub API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdatePlanDto": {
        "type": "object",
        "properties": {
          "officeLocationId": {
            "type": "string",
            "description": "Office Location ID",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "name": {
            "type": "string",
            "description": "Plan name",
            "example": "Basic Plan"
          },
          "slug": {
            "type": "string",
            "description": "Plan slug",
            "example": "basic-plan"
          },
          "description": {
            "type": "string",
            "description": "Plan description",
            "example": "Basic plan with essential features"
          },
          "imageUrl": {
            "type": "string",
            "description": "Plan image URL",
            "example": "https://example.com/plan-image.png"
          },
          "isActive": {
            "type": "boolean",
            "description": "Is plan active",
            "example": true,
            "default": true
          },
          "displayFeatures": {
            "type": "object",
            "description": "Display features for UI (JSON format)",
            "example": {
              "highlights": [
                "Feature 1",
                "Feature 2"
              ]
            }
          },
          "showOnMarketplace": {
            "type": "boolean",
            "description": "Show on marketplace",
            "example": true,
            "default": true
          },
          "isDeleted": {
            "type": "boolean",
            "description": "Is plan deleted",
            "example": false,
            "default": false
          }
        }
      },
      "UpdatePlanPriceDto": {
        "type": "object",
        "properties": {
          "planId": {
            "type": "string",
            "description": "Plan ID",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "billingCycle": {
            "type": "string",
            "description": "Billing cycle",
            "enum": [
              "MONTHLY",
              "YEARLY",
              "WEEKLY",
              "QUARTERLY",
              "ONE_TIME"
            ],
            "example": "MONTHLY",
            "default": "MONTHLY"
          },
          "amount": {
            "type": "number",
            "description": "Price amount in cents",
            "example": 2999,
            "minimum": 0
          },
          "currency": {
            "type": "string",
            "description": "Currency code",
            "example": "USD"
          },
          "description": {
            "type": "string",
            "description": "Price description",
            "example": "Monthly subscription fee"
          },
          "stripePriceId": {
            "type": "string",
            "description": "Stripe Price ID",
            "example": "price_1234567890"
          },
          "isActive": {
            "type": "boolean",
            "description": "Is price active",
            "example": true,
            "default": true
          },
          "isDeleted": {
            "type": "boolean",
            "description": "Is price deleted",
            "example": false
          }
        }
      },
...
      "PlanFeatureResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Plan feature ID",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "planId": {
            "type": "string",
            "description": "Plan ID",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "featureId": {
            "type": "string",
            "description": "Feature ID",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "includedLimit": {
            "type": "number",
            "description": "Included limit",
            "example": 10
          },
          "unitPrice": {
            "type": "number",
            "description": "Unit price in cents",
            "example": 299
          },
          "resetCycle": {
            "type": "string",
            "description": "Reset cycle",
            "enum": [
              "MONTHLY",
              "YEARLY",
              "WEEKLY",
              "QUARTERLY",
              "ONE_TIME",
              "DAILY",
              "NO_RESET"
            ],
            "example": "MONTHLY"
          },
          "isActive": {
            "type": "boolean",
            "description": "Is feature active",
            "example": true
          },
          "displayOrder": {
            "type": "number",
            "description": "Display order",
            "example": 1
          },
          "showOnList": {
            "type": "boolean",
            "description": "Show on list",
            "example": true
          },
          "isDeleted": {
            "type": "boolean",
            "description": "Is feature deleted",
            "example": false
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Created at",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Updated at",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Deleted at",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "plan": {
            "type": "object",
            "description": "Plan details"
          },
          "feature": {
            "type": "object",
            "description": "Feature details"
          }
        },
        "required": [
          "id",
          "planId",
          "featureId",
          "resetCycle",
          "isActive",
          "showOnList",
          "isDeleted",
          "createdAt",
          "updatedAt"
        ]
      },
      "PlanPriceResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Plan price ID",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "planId": {
            "type": "string",
            "description": "Plan ID",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "billingCycle": {
            "type": "string",
            "description": "Billing cycle",
            "enum": [
              "MONTHLY",
              "YEARLY",
              "WEEKLY",
              "QUARTERLY",
              "ONE_TIME"
            ],
            "example": "MONTHLY"
          },
          "amount": {
            "type": "number",
            "description": "Price amount in cents",
            "example": 2999
          },
          "currency": {
            "type": "string",
            "description": "Currency code",
            "example": "USD"
          },
          "description": {
            "type": "string",
            "description": "Price description",
            "example": "Monthly subscription fee"
          },
          "stripePriceId": {
            "type": "string",
            "description": "Stripe Price ID",
            "example": "price_1234567890"
          },
          "isActive": {
            "type": "boolean",
            "description": "Is price active",
            "example": true
          },
          "isDeleted": {
            "type": "boolean",
            "description": "Is price deleted",
            "example": false
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Created at",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Updated at",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Deleted at",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "plan": {
            "type": "object",
            "description": "Plan details"
          },
          "mailboxes": {
            "description": "Mailboxes using this price",
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        },
        "required": [
          "id",
          "planId",
          "billingCycle",
          "amount",
          "currency",
          "isActive",
          "isDeleted",
          "createdAt",
          "updatedAt"
        ]
      },
      "PlanAddonResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Plan addon ID",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "planId": {
            "type": "string",
            "description": "Plan ID",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "productId": {
            "type": "string",
            "description": "Product ID",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "productPriceId": {
            "type": "string",
            "description": "Product Price ID",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "displayOrder": {
            "type": "number",
            "description": "Display order",
            "example": 1
          },
          "isActive": {
            "type": "boolean",
            "description": "Is addon active",
            "example": true
          },
          "isDeleted": {
            "type": "boolean",
            "description": "Is addon deleted",
            "example": false
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Created at",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Updated at",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Deleted at",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "plan": {
            "type": "object",
            "description": "Plan details"
          },
          "product": {
            "type": "object",
            "description": "Product details"
          },
          "prices": {
            "type": "object",
            "description": "Price details"
          }
        },
        "required": [
          "id",
          "planId",
          "productId",
          "productPriceId",
          "displayOrder",
          "isActive",
          "isDeleted",
          "createdAt",
          "updatedAt"
        ]
      },
      "PlanResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Plan ID",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "officeLocationId": {
            "type": "string",
            "description": "Office location ID",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "name": {
            "type": "string",
            "description": "Plan name",
            "example": "Basic Plan"
          },
          "slug": {
            "type": "string",
            "description": "Plan slug",
            "example": "basic-plan"
          },
          "description": {
            "type": "string",
            "description": "Plan description",
            "example": "Basic plan with essential features"
          },
          "imageUrl": {
            "type": "string",
            "description": "Plan image URL",
            "example": "https://example.com/plan-image.png"
          },
          "isActive": {
            "type": "boolean",
            "description": "Is plan active",
            "example": true
          },
          "isDeleted": {
            "type": "boolean",
            "description": "Is plan deleted",
            "example": false
          },
          "showOnMarketplace": {
            "type": "boolean",
            "description": "Show on marketplace",
            "example": true
          },
          "displayFeatures": {
            "type": "object",
            "description": "Display features for UI"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Created at",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Updated at",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Deleted at",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "officeLocation": {
            "type": "object",
            "description": "Office location details"
          },
          "features": {
            "description": "Plan features",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlanFeatureResponseDto"
            }
          },
          "prices": {
            "description": "Plan prices",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlanPriceResponseDto"
            }
          },
          "addons": {
            "description": "Plan addons",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlanAddonResponseDto"
            }
          },
          "mailboxes": {
            "description": "Mailboxes using this plan",
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        },
        "required": [
          "id",
          "officeLocationId",
          "name",
          "slug",
          "isActive",
          "isDeleted",
          "showOnMarketplace",
          "createdAt",
          "updatedAt"
        ]
      },
...
    }
  }
}